+++ lib/phoenix/router.ex
@@ -90,8 +90,6 @@ defmodule Phoenix.Router do
       defmodule MyApp.Router do
         use Phoenix.Router
 
-        pipe_through :browser
-
         resources "/pages", PageController, only: [:show]
         resources "/users", UserController, except: [:delete]
       end
@@ -150,7 +148,6 @@ defmodule Phoenix.Router do
   Note that router pipelines are only invoked after a route is found.
   No plug is invoked in case no matches were found.
 
-
   ### Channels
 
   Channels allow you to route pubsub events to channel handlers in your application.
@@ -170,13 +167,35 @@ defmodule Phoenix.Router do
 
   alias Phoenix.Router.Resource
   alias Phoenix.Router.Scope
+  alias Phoenix.Router.Route
   alias Phoenix.Router.Helpers
 
   @http_methods [:get, :post, :put, :patch, :delete, :options, :connect, :trace, :head]
 
   @doc false
-  defmacro __using__(_opts) do
+  defmacro __using__(opts) do
     quote do
+      opts = unquote(opts)
+      @pubsub_server opts[:pubsub_server] ||
+        Phoenix.Naming.base_concat(__MODULE__, "PubSub")
+
+      def __pubsub_server__, do: @pubsub_server
+
+      def broadcast_from(from, topic, event, msg) when is_map(msg) do
+        Phoenix.Channel.broadcast_from(@pubsub_server, from, topic, event, msg)
+      end
+      def broadcast_from!(from, topic, event, msg) when is_map(msg) do
+        Phoenix.Channel.broadcast_from!(@pubsub_server, from, topic, event, msg)
+      end
+
+      def broadcast(topic, event, msg) do
+        Phoenix.Channel.broadcast(@pubsub_server, topic, event, msg)
+      end
+
+      def broadcast!(topic, event, msg) do
+        Phoenix.Channel.broadcast!(@pubsub_server, topic, event, msg)
+      end
+
       unquote(prelude())
       unquote(plug())
     end
@@ -246,18 +265,22 @@ defmodule Phoenix.Router do
   @doc false
   defmacro __before_compile__(env) do
     routes   = env.module |> Module.get_attribute(:phoenix_routes) |> Enum.reverse
-    chan_ast = env.module |> Module.get_attribute(:phoenix_channels) |> Helpers.defchannels
+    channels = env.module |> Module.get_attribute(:phoenix_channels) |> Helpers.defchannels
 
     Helpers.define(env, routes)
 
     quote do
       @doc false
-      def __routes__, do: unquote(Macro.escape(routes))
+      def __routes__,  do: unquote(Macro.escape(routes))
+
+      @doc false
+      def __helpers__, do: __MODULE__.Helpers
 
       defp match(conn, _method, _path_info, _host) do
         raise NoRouteError, conn: conn, router: __MODULE__
       end
-      unquote(chan_ast)
+
+      unquote(channels)
     end
   end
 
@@ -280,12 +303,16 @@ defmodule Phoenix.Router do
 
       defp match(var!(conn), unquote(route.verb), unquote(route.path_segments),
                  unquote(route.host_segments)) do
+        unquote(Route.maybe_merge(:private, route.private))
+
         var!(conn) =
-          Plug.Conn.put_private(var!(conn), :phoenix_route, fn conn ->
-            update_in(conn.params, &Map.merge(&1, unquote(parts)))
-            |> unquote(route.controller).call(unquote(route.controller).init(unquote(route.action)))
-          end)
+          update_in(var!(conn).params, &Map.merge(&1, unquote(parts)))
           |> Plug.Conn.put_private(:phoenix_pipelines, unquote(route.pipe_through))
+          |> Plug.Conn.put_private(:phoenix_route, fn conn ->
+              opts = unquote(route.controller).init(unquote(route.action))
+              unquote(route.controller).call(conn, opts)
+             end)
+
         unquote(route.pipe_segments)
       end
     end
@@ -415,6 +442,53 @@ defmodule Phoenix.Router do
     add_resources path, controller, [], do: nil
   end
 
+  @doc """
+  Defines "RESTful" routes for a resource that client's lookup without referencing an ID.
+
+  The given definition:
+
+      resource "/account", UserController
+
+  will include routes to the following actions:
+
+    * `GET /account` => `:show`
+    * `GET /account/new` => `:new`
+    * `POST /account` => `:create`
+    * `GET /account/edit` => `:edit`
+    * `PATCH /account` => `:update`
+    * `PUT /account` => `:update`
+    * `DELETE /account` => `:delete`
+
+  ## Options
+
+  This macro accepts the same options as `resources/4`
+
+  """
+  defmacro resource(path, controller, opts, do: nested_context) do
+    add_resource path, controller, opts, do: nested_context
+  end
+
+  @doc """
+  See `resource/4`.
+  """
+  defmacro resource(path, controller, do: nested_context) do
+    add_resource path, controller, [], do: nested_context
+  end
+
+  @doc """
+  See `resource/4`.
+  """
+  defmacro resource(path, controller, opts) do
+    add_resource path, controller, opts, do: nil
+  end
+
+  @doc """
+  See `resource/4`.
+  """
+  defmacro resource(path, controller) do
+    add_resource path, controller, [], do: nil
+  end
+
   defp add_resources(path, controller, options, do: context) do
     quote do
       resource = Resource.build(unquote(path), unquote(controller), unquote(options))
@@ -422,7 +496,7 @@ defmodule Phoenix.Router do
       parm = resource.param
       path = resource.path
       ctrl = resource.controller
-      opts = [as: resource.as]
+      opts = resource.route
 
       Enum.each resource.actions, fn action ->
         case action do
@@ -434,7 +508,35 @@ defmodule Phoenix.Router do
           :delete  -> delete "#{path}/:#{parm}",      ctrl, :delete, opts
           :update  ->
             patch "#{path}/:#{parm}", ctrl, :update, opts
-            put   "#{path}/:#{parm}", ctrl, :update, as: nil
+            put   "#{path}/:#{parm}", ctrl, :update, Keyword.put(opts, :as, nil)
+        end
+      end
+
+      scope resource.member do
+        unquote(context)
+      end
+    end
+  end
+
+  defp add_resource(path, controller, options, do: context) do
+    quote do
+      opts     = Keyword.merge(unquote(options), singular: true)
+      resource = Resource.build(unquote(path), unquote(controller), opts)
+
+      path = resource.path
+      ctrl = resource.controller
+      opts = resource.route
+
+      Enum.each resource.actions, fn action ->
+        case action do
+          :show    -> get    "#{path}",      ctrl, :show, opts
+          :new     -> get    "#{path}/new",  ctrl, :new, opts
+          :edit    -> get    "#{path}/edit", ctrl, :edit, opts
+          :create  -> post   "#{path}",      ctrl, :create, opts
+          :delete  -> delete "#{path}",      ctrl, :delete, opts
+          :update  ->
+            patch "#{path}", ctrl, :update, opts
+            put   "#{path}", ctrl, :update, Keyword.put(opts, :as, nil)
         end
       end
 
@@ -466,6 +568,7 @@ defmodule Phoenix.Router do
     * `:alias` - an alias (atom) containing the controller scope
     * `:host` - a string containing the host scope, or prefix host scope, ie
                 `"foo.bar.com"`, `"foo."`
+    * `:private` - a map of private data to merge into the connection when a route matches
 
   """
   defmacro scope(options, do: context) do
@@ -524,19 +627,25 @@ defmodule Phoenix.Router do
   end
 
   @doc """
-  Defines a socket mount-point for channel definitions. By default, the
-  given path is a websocket upgrade endpoint, with Long-polling fallback.
-  The transports can be configured with the socket options or on each individual
-  channel.
+  Defines a socket mount-point for channel definitions.
 
-    * `mount` - The string path for the websocket upgrade, ie "/ws"
-    * `opts` - The optional keyword list of options
-      * `via` - The optional transport modules to apply to all channels in the block,
-                ie: `[Phoenix.Transports.WebSocket]`
-      *`as` - The optional named route helper function, ie :socket
-      *`alias` - The optional alias to apply to all channel modules, ie: MyApp.
-                 Alternatively, you can pass an alias as a standalone second argument
-                 to `socket` to apply the alias, similar to `scope`.
+  By default, the given path is a websocket upgrade endpoint,
+  with long-polling fallback. The transports can be configured
+  with the socket options or on each individual channel.
+
+  It expects the `mount` path as a string and a keyword list
+  of options.
+
+  ## Options
+
+    * `:via` - the optional transport modules to apply to all
+      channels in the block, ie: `[Phoenix.Transports.WebSocket]`
+
+    * `:as` - the optional named route helper function, ie `:socket`
+
+    * `:alias` - the optional alias to apply to all channel modules,
+      ie: `MyApp`. Alternatively, you can pass an alias as a standalone
+      second argument to apply the alias, similar to `scope/2`.
 
   ## Examples
 
@@ -587,14 +696,18 @@ defmodule Phoenix.Router do
       @phoenix_channel_alias nil
     end
   end
+
   @doc """
-  Defines a channel matching the given topic and transports
+  Defines a channel matching the given topic and transports.
 
     * `topic_pattern` - The string pattern, ie "rooms:*", "users:*", "system"
     * `module` - The channel module handler, ie `MyApp.RoomChannel`
-    * `opts` - The optional list of options. Available options are:
-      * `via` - The transport adapters to accept on this channel.
-                Defaults `[Phoenix.Transports.WebSocket, Phoenix.Transports.LongPoller]`
+    * `opts` - The optional list of options, see below
+
+  ## Options
+
+    * `:via` - the transport adapters to accept on this channel.
+      Defaults `[Phoenix.Transports.WebSocket, Phoenix.Transports.LongPoller]`
 
   ## Examples