+++ lib/phoenix/transports/long_poller.ex
@@ -6,21 +6,7 @@ defmodule Phoenix.Transports.LongPoller do
   alias Phoenix.Socket.Message
   alias Phoenix.Transports.LongPoller
 
-  # TODO: If we are going to decouple this from the router,
-  # we need to plug the parameter parser and use its own
-  # session thing
-
-  plug :fetch_session
   plug :action
-
-  @doc """
-  Starts `Phoenix.LongPoller.Server` and stores pid in session. This action must
-  be called first before sending requests to `poll` or `publish`
-  """
-  def open(conn, _) do
-    {conn, _server_pid} = start_session(conn)
-    send_resp(conn, :ok, "")
-  end
 
   @doc """
   Listens for `%Phoenix.Socket.Message{}`'s from `Phoenix.LongPoller.Server`.
@@ -31,22 +17,29 @@ defmodule Phoenix.Transports.LongPoller do
   """
   def poll(conn, _params) do
     case resume_session(conn) do
-      {:ok, conn, server_pid}     -> listen(conn, server_pid)
-      {:error, conn, :terminated} -> send_resp(conn, :gone, "")
+      {:ok, conn, priv_topic}     -> listen(conn, priv_topic)
+      {:error, conn, :terminated} ->
+        {conn, priv_topic, sig, _server_pid} = start_session(conn)
+
+        conn
+        |> put_status(:gone)
+        |> json(%{token: priv_topic, sig: sig})
     end
   end
-  defp listen(conn, server_pid) do
-    timeout_ms = timeout_window_ms(conn)
-    set_active_listener(server_pid, self)
+  defp listen(conn, priv_topic) do
+    ref = :erlang.make_ref()
+    :ok = broadcast_from(conn, priv_topic, {:flush, ref})
 
     receive do
-      {:messages, msgs} ->
-        ack(server_pid, msgs)
-        json(conn, msgs)
+      {:messages, msgs, ^ref} ->
+        :ok = ack(conn, priv_topic, msgs)
+        json(conn, %{messages: msgs, token: conn.params["token"], sig: conn.params["sig"]})
     after
-      timeout_ms ->
-        ack(server_pid, [])
-        send_resp(conn, :no_content, "")
+      timeout_window_ms(conn) ->
+        :ok = ack(conn, priv_topic, [])
+        conn
+        |> put_status(:no_content)
+        |> json(%{token: conn.params["token"], sig: conn.params["sig"]})
     end
   end
 
@@ -58,17 +51,17 @@ defmodule Phoenix.Transports.LongPoller do
   """
   def publish(conn, message) do
     case resume_session(conn) do
-      {:ok, conn, server_pid}     -> dispatch_publish(conn, message, server_pid)
-      {:error, conn, :terminated} -> send_resp(conn, :gone, "")
+      {:ok, conn, priv_topic}     -> dispatch_publish(conn, message, priv_topic)
+      {:error, conn, :terminated} -> conn |> put_status(:gone) |> json(%{})
     end
   end
 
-  defp dispatch_publish(conn, message, server_pid) do
+  defp dispatch_publish(conn, message, priv_topic) do
     msg = Message.from_map!(message)
 
-    case dispatch(server_pid, msg) do
-      {:ok, _socket}             -> send_resp(conn, :ok, "")
-      {:error, _socket, _reason} -> send_resp(conn, :unauthorized, "")
+    case dispatch(conn, priv_topic, msg) do
+      :ok               -> send_resp(conn, :ok, "")
+      {:error, _reason} -> send_resp(conn, :unauthorized, "")
     end
   end
 
@@ -79,75 +72,104 @@ defmodule Phoenix.Transports.LongPoller do
   """
   def start_session(conn) do
     router = router_module(conn)
-    child  = [router, timeout_window_ms(conn)]
+    priv_topic =
+      "phx:lp:"
+      |> Kernel.<>(Base.encode64(:crypto.strong_rand_bytes(16)))
+      |> Kernel.<>(:os.timestamp() |> Tuple.to_list |> Enum.join(""))
+
+    child = [router, timeout_window_ms(conn), priv_topic, pubsub_server(conn)]
     {:ok, server_pid} = Supervisor.start_child(LongPoller.Supervisor, child)
-    conn = put_session_with_salt(conn, server_pid)
 
-    {conn, server_pid}
-  end
-
-  # Serialized longpoll server pid into session and harden with random salt
-  defp put_session_with_salt(conn, server_pid) do
-    key = session_key(conn)
-    conn
-    |> put_session(key, :erlang.term_to_binary(server_pid))
-    |> put_session("#{key}_salt", :crypto.strong_rand_bytes(16) |> Base.encode64)
+    {conn, priv_topic, sign(conn, priv_topic), server_pid}
   end
 
   @doc """
   Finds the `Phoenix.LongPoller.Server` server from the session
   """
   def resume_session(conn) do
-    case longpoll_pid(conn) do
-      {:ok, pid}            -> {:ok, conn, pid}
-      :nopid                -> {:error, conn, :terminated}
+    case verify_longpoll_topic(conn) do
+      {:ok, priv_topic}     -> {:ok, conn, priv_topic}
+      :notopic              -> {:error, conn, :terminated}
       {:error, :terminated} -> {:error, conn, :terminated}
     end
   end
 
   @doc """
-  Retrieves the serialized `Phoenix.LongPoller.Server` pid from the session
+  Retrieves the serialized `Phoenix.LongPoller.Server` pid from the encrypted token
   """
-  def longpoll_pid(conn) do
-    case get_session(conn, session_key(conn)) do
-      nil -> :nopid
-      bin ->
-        pid = :erlang.binary_to_term(bin)
-        if Process.alive?(pid) do
-          {:ok, pid}
-        else
-          {:error, :terminated}
+  def verify_longpoll_topic(%Plug.Conn{params: %{"token" => token, "sig" => sig}} = conn) do
+    case verify(conn, token, sig) do
+      {:ok, priv_topic} ->
+        ref = :erlang.make_ref()
+        :ok = subscribe(conn, priv_topic)
+        :ok = broadcast_from(conn, priv_topic, {:subscribe, ref})
+        receive do
+          {:ok, :subscribe, ^ref} -> {:ok, priv_topic}
+        after
+          pubsub_timeout_ms(conn)  -> {:error, :terminated}
         end
+
+      _ -> :notopic
     end
   end
+  def verify_longpoll_topic(_conn), do: :notopic
 
   @doc """
   Ack's a list of `%Phoenix.Socket.Messages{}`'s back to the `Phoenix.LongPoller.Server`
 
   To be called after buffered messages have been relayed to client
   """
-  def ack(server_pid, messages) do
-    :ok = GenServer.call(server_pid, {:ack, messages})
-  end
-
-  defp session_key(conn), do: "#{router_module(conn)}_longpoll_pid"
-
-  @doc """
-  Sets the active listener process. Called by polling `conn`
-  """
-  def set_active_listener(server_pid, listener_pid) do
-    GenServer.call(server_pid, {:set_active_listener, listener_pid})
+  def ack(conn, priv_topic, messages) do
+    ref = :erlang.make_ref()
+    :ok = broadcast_from(conn, priv_topic, {:ack, messages, ref})
+    receive do
+      {:ok, :ack, ^ref} -> :ok
+    after
+      pubsub_timeout_ms(conn) -> :error
+    end
   end
 
   @doc """
   Dispatches deserialized `%Phoenix.Socket.Message{}` from client to
   `Phoenix.LongPoller.Server`
   """
-  def dispatch(server_pid, message) do
-    GenServer.call(server_pid, {:dispatch, message})
+  def dispatch(conn, priv_topic, msg) do
+    ref = :erlang.make_ref()
+    :ok = broadcast_from(conn, priv_topic, {:dispatch, msg, ref})
+    receive do
+      {:ok, :dispatch, ^ref}            -> :ok
+      {:error, :dispatch, reason, ^ref} -> {:error, reason}
+    after
+      pubsub_timeout_ms(conn) -> {:error, :timeout}
+    end
   end
 
   defp timeout_window_ms(conn) do
     get_in endpoint_module(conn).config(:transports), [:longpoller_window_ms]
+  end
+
+  defp pubsub_timeout_ms(conn) do
+    get_in endpoint_module(conn).config(:transports), [:longpoller_pubsub_timeout_ms]
+  end
+
+  defp pubsub_server(conn), do: router_module(conn).__pubsub_server__()
+
+  defp subscribe(conn, priv_topic) do
+    Phoenix.PubSub.subscribe(pubsub_server(conn), self, priv_topic)
+  end
+
+  defp broadcast_from(conn, priv_topic, msg) do
+    Phoenix.PubSub.broadcast_from(pubsub_server(conn), self, priv_topic, msg)
+  end
+
+  defp sign(conn, priv_topic) do
+    Plug.Crypto.MessageVerifier.sign(priv_topic, conn.secret_key_base)
+  end
+
+  defp verify(conn, priv_topic, sig) do
+    case Plug.Crypto.MessageVerifier.verify(sig, conn.secret_key_base) do
+      {:ok, ^priv_topic} -> {:ok, priv_topic}
+      _ -> :error
+    end
   end
 end