+++ test/phoenix/router/helpers_test.exs
@@ -7,7 +7,7 @@ defmodule Phoenix.Router.HelpersTest do
   ## Unit tests
 
   test "defhelper with :identifiers" do
-    route = build("GET", "/foo/:bar", nil, Hello, :world, "hello_world", [])
+    route = build("GET", "/foo/:bar", nil, Hello, :world, "hello_world")
 
     assert extract_defhelper(route, 0) == String.strip """
     def(hello_world_path(conn_or_endpoint, :world, bar)) do
@@ -23,7 +23,7 @@ defmodule Phoenix.Router.HelpersTest do
   end
 
   test "defhelper with *identifiers" do
-    route = build("GET", "/foo/*bar", nil, Hello, :world, "hello_world", [])
+    route = build("GET", "/foo/*bar", nil, Hello, :world, "hello_world")
 
     assert extract_defhelper(route, 0) == String.strip """
     def(hello_world_path(conn_or_endpoint, :world, bar)) do
@@ -38,8 +38,8 @@ defmodule Phoenix.Router.HelpersTest do
     """
   end
 
-  defp build(verb, path, host, controller, action, helper, pipe_through) do
-    Phoenix.Router.Route.build(verb, path, host, controller, action, helper, pipe_through)
+  defp build(verb, path, host, controller, action, helper) do
+    Phoenix.Router.Route.build(verb, path, host, controller, action, helper, [], %{})
   end
 
   defp extract_defhelper(route, pos) do
@@ -68,6 +68,10 @@ defmodule Phoenix.Router.HelpersTest do
 
     resources "/files", FileController
 
+    resource "/account", UserController, as: :account do
+      resource "/page", PagesController, as: :page, only: [:show]
+    end
+
     scope "/admin", alias: Admin do
       resources "/messages", MessageController
     end
@@ -80,6 +84,10 @@ defmodule Phoenix.Router.HelpersTest do
   end
 
   alias Router.Helpers
+
+  test "defines a __helpers__ function" do
+    assert Router.__helpers__ == Router.Helpers
+  end
 
   test "top-level named route" do
     assert Helpers.post_path(__MODULE__, :show, 5) == "/posts/5"
@@ -168,6 +176,24 @@ defmodule Phoenix.Router.HelpersTest do
     assert Helpers.file_path(__MODULE__, :show, 123) == "/files/123"
     assert Helpers.file_path(__MODULE__, :new, []) == "/files/new"
     assert Helpers.file_path(__MODULE__, :new) == "/files/new"
+  end
+
+  test "resource generates named routes for :show, :edit, :new, :update, :delete" do
+    assert Helpers.account_path(__MODULE__, :show, []) == "/account"
+    assert Helpers.account_path(__MODULE__, :show) == "/account"
+    assert Helpers.account_path(__MODULE__, :edit, []) == "/account/edit"
+    assert Helpers.account_path(__MODULE__, :edit) == "/account/edit"
+    assert Helpers.account_path(__MODULE__, :new, []) == "/account/new"
+    assert Helpers.account_path(__MODULE__, :new) == "/account/new"
+    assert Helpers.account_path(__MODULE__, :update, []) == "/account"
+    assert Helpers.account_path(__MODULE__, :update) == "/account"
+    assert Helpers.account_path(__MODULE__, :delete, []) == "/account"
+    assert Helpers.account_path(__MODULE__, :delete) == "/account"
+  end
+
+  test "2-Level nested resource generates nested named routes for :show" do
+    assert Helpers.account_page_path(__MODULE__, :show, []) == "/account/page"
+    assert Helpers.account_page_path(__MODULE__, :show) == "/account/page"
   end
 
   test "scoped route helpers generated named routes with :path, and :alias options" do