+++ test/phoenix/router/route_test.exs
@@ -4,7 +4,7 @@ defmodule Phoenix.Router.RouteTest do
   import Phoenix.Router.Route
 
   test "builds a route based on verb, path, controller, action and helper" do
-    route = build("GET", "/foo/:bar", nil, Hello, :world, "hello_world", [:foo, :bar])
+    route = build("GET", "/foo/:bar", nil, Hello, :world, "hello_world", [:foo, :bar], %{foo: "bar"})
     assert route.verb == "GET"
     assert route.path == "/foo/:bar"
     assert route.host == nil
@@ -16,13 +16,14 @@ defmodule Phoenix.Router.RouteTest do
     assert route.path_segments == ["foo", {:bar, [], nil}]
     assert Macro.to_string(route.host_segments) == "_"
     assert Macro.to_string(route.pipe_segments) == "bar(foo(var!(conn), []), [])"
+    assert route.private == %{foo: "bar"}
   end
 
   test "builds a route based on the host" do
-    route = build("GET", "/foo/:bar", "foo.", Hello, :world, "hello_world", [])
+    route = build("GET", "/foo/:bar", "foo.", Hello, :world, "hello_world", [], %{foo: "bar"})
     assert Macro.to_string(route.host_segments) == "\"foo.\" <> _"
 
-    route = build("GET", "/foo/:bar", "foo.com", Hello, :world, "hello_world", [])
+    route = build("GET", "/foo/:bar", "foo.com", Hello, :world, "hello_world", [], %{foo: "bar"})
     assert Macro.to_string(route.host_segments) == "\"foo.com\""
   end
 end